ISSUE #711 Changes to alert settings [CHANGED] alert range -> Float

+++ app/src/main/java/com/ruuvi/station/alarm/domain/AlarmItemState.kt
@@ -7,8 +7,8 @@ data class AlarmItemState(
     val sensorId: String,
     var type: AlarmType,
     var isEnabled: Boolean,
-    var low: Int,
-    var high: Int,
+    var low: Float,
+    var high: Float,
     var customDescription: String = "",
     var mutedTill: Date? = null,
 ) {
@@ -16,8 +16,8 @@ data class AlarmItemState(
         sensorId = alarm.ruuviTagId,
         type = alarm.alarmType,
         isEnabled = alarm.enabled,
-        low = alarm.low,
-        high = alarm.high,
+        low = alarm.low.toFloat(),
+        high = alarm.high.toFloat(),
         customDescription = alarm.customDescription,
         mutedTill = alarm.mutedTill
     ) {
@@ -28,15 +28,15 @@ data class AlarmItemState(
         sensorId = sensorId,
         type = alarmType,
         isEnabled = false,
-        low = alarmType.possibleRange.first,
-        high = alarmType.possibleRange.last
+        low = alarmType.possibleRange.first.toFloat(),
+        high = alarmType.possibleRange.last.toFloat()
     )
 
     fun getPossibleRange() = type.possibleRange
 
     fun normalizeValues() {
-        val min = getPossibleRange().first
-        val max = getPossibleRange().last
+        val min = getPossibleRange().first.toFloat()
+        val max = getPossibleRange().last.toFloat()
         if (low < min) low = min
         if (low >= max) low = max - 1
         if (high > max) high = max
